move tests from test_packet.rs to packet.rs, along with the other unittests
