+++ Cargo.toml
@@ -28,10 +28,6 @@ name = "tftp_server_bin"
 path = "src/bin.rs"
 
 [[test]]
-name = "tftp-packet-tests"
-path = "tests/test_packet.rs"
-
-[[test]]
 name = "tftp-server-tests"
 path = "tests/test_server.rs"
 harness = false
